:recycle:  refactor: editor darkMode 추가 설정 훅으로 분리
